Revert "Merge branches 'API_update' and 'main' of github.com:sdm4fzi/prodsys"
This reverts commit e88e1f0820d5af0d4efcd58fd41a257a45250dd4, reversing
changes made to 182f3d1a3bc94f55efbd8693f35803db468c7478.

+++ prodsys/simulation/sim.py
@@ -63,7 +63,7 @@ class Environment(core.Environment):
         self.pbar: Any = None
         self.last_update = 0
 
-    def run(self, time_range: float):
+    def run(self, time_range:int):
         """
         Runs the simulation for a given time range.
